Consolidate `connect` logic
