+++ lib/client.js
@@ -1,24 +1,7 @@
 'use strict';
 
-const net = require('net');
 const out = require('./out');
-const portfile = require('./portfile');
-
-function connect(callback) {
-  portfile.read((config) => {
-    if (!config) {
-      callback('Not running');
-      return;
-    }
-    const socket = net.connect(config.port, '127.0.0.1', () => {
-      callback(null, socket, config.token);
-    });
-    socket.once('error', () => {
-      process.exitCode = 1;
-      callback('Could not connect');
-    });
-  });
-}
+const connect = require('./connect');
 
 exports.stop = function (callback) {
   connect((err, socket, token) => {
