+++ lib/connect.js
@@ -1,0 +1,20 @@
+'use strict';
+
+const net = require('net');
+const portfile = require('./portfile');
+
+module.exports = function (callback) {
+  portfile.read((config) => {
+    if (!config) {
+      callback('Not running');
+      return;
+    }
+    const socket = net.connect(config.port, '127.0.0.1', () => {
+      callback(null, socket, config.token);
+    });
+    socket.once('error', () => {
+      process.exitCode = 1;
+      callback('Could not connect');
+    });
+  });
+};
