+++ lib/launcher.js
@@ -1,26 +1,10 @@
 'use strict';
 
-const net = require('net');
 const out = require('./out');
-const portfile = require('./portfile');
-
-function check(callback) {
-  portfile.read((config) => {
-    if (!config) {
-      callback('Not running');
-      return;
-    }
-    const socket = net.connect(config.port, '127.0.0.1', () => {
-      callback(null, socket, config.token);
-    });
-    socket.once('error', () => {
-      callback('Could not connect');
-    });
-  });
-}
+const connect = require('./connect');
 
 function wait(callback) {
-  check((err, socket, token) => {
+  connect((err, socket, token) => {
     if (err) {
       setTimeout(() => {
         wait(callback);
@@ -54,7 +38,7 @@ function launch(callback) {
 }
 
 exports.launch = function (callback) {
-  check((err, socket) => {
+  connect((err, socket) => {
     if (err) {
       launch(callback);
     } else {
